feat: improve math for single asset join (#1115)
* feat: improve math for single asset join

+++ x/common/common.go
@@ -38,7 +38,7 @@ var (
 	// Precision for int representation in sdk.Int objects
 	Precision = int64(1_000_000)
 
-	BigIntPrecision = 18
+	BigIntPrecision = int64(18)
 )
 
 //-----------------------------------------------------------------------------
